PPS: reduce chance of lost PPS edges
The PPS edge might happen while time_pps_fetch() is not running and
the PPS edge could be lost which would lead to process_pps() waiting
for the full idle_sleep time (default 25ms). Reduce the chance of this
happening by first doing a non-blocking check if an edge occured while
time_pps_fetch() was not running and only blocking if this was not the
case. This still has the very small potential of an edge occuring
between the non-blocking and the blocking call to time_pps_fetch()
which needs to be solved later.